refactor: Merged part of landing page
