+++ src/components/LandingPage/Sections/MainSection/main-section.module.scss.d.ts
@@ -1,1 +1,3 @@
 export const mainSection: string;
+export const title: string;
+export const subTitle: string;
