+++ src/pages/index.tsx
@@ -2,6 +2,8 @@ import React from 'react';
 
 import MainSection from '../components/LandingPage/Sections/MainSection/main-section';
 import Layout from '../templates/Layout/layout';
+import Link from '../components/Link';
+import Hero from '../components/hero';
 
 // should be reaplaced with GraphQL data
 const title = 'Node.js Official website';
@@ -10,10 +12,6 @@ const content = {
   section1_1: 'The power of JavaScript minus the browser',
   section1_2: `Tacos raw denim palo santo, squid 90's occupy sustainable leggings locavore before they sold out chambray gastropub synth. Cornhole lomo dreamcatcher celiac enamel pin, yr lyft master cleanse meggings.`,
 };
-import { Link } from 'gatsby';
-import Hero from '../components/hero';
-import Layout from '../components/layout';
-import '../util/konami';
 
 import '../styles/index.css';
 
@@ -51,103 +49,92 @@ const NodeFeature = ({ img, featureText }: Props) => {
   );
 };
 
+interface Props {
+  img: string;
+  featureText: string;
+}
+
 export default function Index(): JSX.Element {
-  const title = 'The power of JavaScript minus the browser';
-  const subTitle =
-    'Node.js is a free, open-sourced, cross-platform JavaScript run-time environment that lets developers write command line tools and server-side scripts outside of a browser.';
   const description = 'Welcome to Node.js!';
 
-export default () => {
   return (
-    <Layout title={title} description={description}>
-      <MainSection content={content} />
-      <MainSection content={content} />
-      <MainSection content={content} />
-    </Layout>
-  );
-};
     <Layout
       title={title}
       description={description}
-      style={{ overflow: 'hidden' }}
     >
-      <div className="home-page">
-        <Hero title={title} subTitle={subTitle} />
+      <MainSection content={content} />
 
-        <section className="node-demo-container">
-          <div className="node-demo"></div>
-          <img className="leafs-front" src={leafsIllustrationFront} alt="" />
-          <img className="leafs-middle" src={leafsIllustrationMiddle} alt="" />
-          <img className="leafs-back" src={leafsIllustrationBack} alt="" />
-          <img className="dots" src={dotsIllustration} alt="" />
-        </section>
+      <section className="node-demo-container">
+        <div className="node-demo" />
+        <img className="leafs-front" src={leafsIllustrationFront} alt="" />
+        <img className="leafs-middle" src={leafsIllustrationMiddle} alt="" />
+        <img className="leafs-back" src={leafsIllustrationBack} alt="" />
+        <img className="dots" src={dotsIllustration} alt="" />
+      </section>
 
-        <section className="node-features">
-          <NodeFeature img={featureImg} featureText={nodeFeature1} />
-          <NodeFeature img={featureImg} featureText={nodeFeature2} />
-          <NodeFeature img={featureImg} featureText={nodeFeature3} />
-        </section>
+      <section className="node-features">
+        <NodeFeature img={featureImg} featureText={nodeFeature1} />
+        <NodeFeature img={featureImg} featureText={nodeFeature2} />
+        <NodeFeature img={featureImg} featureText={nodeFeature3} />
+      </section>
 
-        <section className="join-node">
-          <h4 className="t-headline">
-            <span className="accent">Join</span> the community
-          </h4>
-          <div className="join-node-form-container">
-            <p className="t-body2">
-              We’ll never share your information and always respect your inbox -
-              quality content only, we promise.{' '}
-            </p>
-            <div style={{ display: `flex`, alignItems: 'flex-start' }}>
-              <input
-                type="email"
-                placeholder="node@nodejs.dev"
-                className="input-subscribe"
-              />
-              <button className="btn-subscribe t-body1" type="button">
-                Subscribe
-              </button>
-            </div>
+      <section className="join-node">
+        <h4 className="t-headline">
+          <span className="accent">Join</span> the community
+        </h4>
+        <div className="join-node-form-container">
+          <p className="t-body2">
+            We’ll never share your information and always respect your inbox -
+            quality content only, we promise.{' '}
+          </p>
+          <div style={{ display: `flex`, alignItems: 'flex-start' }}>
+            <input
+              type="email"
+              placeholder="node@nodejs.dev"
+              className="input-subscribe"
+            />
+            <button className="btn-subscribe t-body1" type="button">
+              Subscribe
+            </button>
           </div>
-        </section>
+        </div>
+      </section>
 
-        <section className="trusted-by">
-          <h4 className="t-headline">
-            Trusted by development teams around the world
-          </h4>
-          <p>Including IBM, LinkedIn, Microsoft, Netflix, and PayPal.</p>
-          <div className="logos-container">
-            <img src={logoImg1} alt="ibm logo" />
-            <img src={logoImg2} alt="linkedin logo" />
-            <img src={logoImg3} alt="microsoft logo" />
-            <img src={logoImg4} alt="netflix logo" />
-            <img src={logoImg5} alt="paypal logo" />
-          </div>
-        </section>
+      <section className="trusted-by">
+        <h4 className="t-headline">
+          Trusted by development teams around the world
+        </h4>
+        <p>Including IBM, LinkedIn, Microsoft, Netflix, and PayPal.</p>
+        <div className="logos-container">
+          <img src={logoImg1} alt="ibm logo" />
+          <img src={logoImg2} alt="linkedin logo" />
+          <img src={logoImg3} alt="microsoft logo" />
+          <img src={logoImg4} alt="netflix logo" />
+          <img src={logoImg5} alt="paypal logo" />
+        </div>
+      </section>
 
-        <section className="get-started-callouts">
-          <Link to="/learn" className="get-started-callout">
-            <img src={GetStartedIllustration1} alt="" />
-            <h5 className="t-headline">A beginner’s guide</h5>
-            <p>
-              Lorem ipsum dolor amet pug vape +1 poke pour-over kitsch tacos
-              meh.
-            </p>
-          </Link>
-
-          <Link to="/docs" className="get-started-callout">
-            <img src={GetStartedIllustration2} alt="" />
-            <h5 className="t-headline">Do even more with Node</h5>
-            <p>
-              Lorem ipsum dolor amet pug vape +1 poke pour-over kitsch tacos
-              meh.
-            </p>
-          </Link>
-        </section>
-
-        <Link to="/learn" className="btn-primary">
-          Get Started
+      <section className="get-started-callouts">
+        <Link to="/learn" className="get-started-callout">
+          <img src={GetStartedIllustration1} alt="" />
+          <h5 className="t-headline">A beginner’s guide</h5>
+          <p>
+            Lorem ipsum dolor amet pug vape +1 poke pour-over kitsch tacos meh.
+          </p>
         </Link>
-      </div>
+
+        <Link to="/docs" className="get-started-callout">
+          <img src={GetStartedIllustration2} alt="" />
+          <h5 className="t-headline">Do even more with Node</h5>
+          <p>
+            Lorem ipsum dolor amet pug vape +1 poke pour-over kitsch tacos meh.
+          </p>
+        </Link>
+      </section>
+
+      <Link to="/learn" className="btn-primary">
+        Get Started
+      </Link>
       <img
         className="pentagon-illustration-big1"
         src={pentagonIllustration1}
@@ -158,12 +145,7 @@ export default () => {
         src={pentagonIllustration2}
         alt=""
       />
-      <div className="double-background"></div>
+      <div className="double-background" />
     </Layout>
   );
-}
-
-interface Props {
-  img: string;
-  featureText: string;
 }
