+++ src/styles/index.css
@@ -1,11 +1,4 @@
-.home-page {
-    max-width: 960px;
-    margin: 0 auto;
-    margin-bottom: var(--space-80);
-    position: relative;
 
-    --section-margin-bottom: 340px;
-}
 
 .pentagon-illustration-big1 {
     position: absolute;
@@ -30,30 +23,6 @@
     z-index: -10;
 }
 
-.home-page-hero {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    text-align: center;
-    padding-top: var(--space-96);
-    margin-bottom: 160px;
-}
-
-.home-page-hero h1, p {
-    margin: 0;
-}
-
-.home-page-hero h1 {
-    max-width: 840px;
-}
-
-.home-page-hero .sub-title {
-    max-width: 780px;
-    color: var(--color-text-secondary);
-    margin: var(--space-32) 0 var(--space-80) 0;
-    font-weight: var(--font-weight-light);
-}
 
 .download-lts-container {
     position: relative;
