+++ src/styles/layout.css
@@ -194,11 +194,6 @@ select:hover, select:active {
   border-color: var(--black5);
 }
 
-main {
-  display: grid;
-  grid-template-columns: 28.8rem auto;
-}
-
 :focus {
   outline: var(--brand3) dotted 2px;
 }
