+++ src/templates/Layout/layout.module.scss
@@ -1,3 +1,8 @@
 .main-container {
-	width: 100%;
+	max-width: 960px;
+    margin: 0 auto;
+    margin-bottom: var(--space-80);
+    position: relative;
+
+    --section-margin-bottom: 340px;
 }
